+++ pytsite/auth_log/__init__.py
@@ -1,12 +1,23 @@
 """Auth Log Package Init.
 """
-from pytsite import events as _events, odm as _odm
-from . import _eh, _odm_model
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-_odm.register_model('auth_log', _odm_model.AuthLog)
-_events.listen('pytsite.auth.login', _eh.auth_login)
-_events.listen('pytsite.auth.logout', _eh.auth_logout)
+
+def __init():
+    from pytsite import events, odm, admin, lang, router
+    from . import _eh, _odm_model
+
+    lang.register_package(__name__)
+
+    odm.register_model('auth_log', _odm_model.AuthLog)
+    events.listen('pytsite.auth.login', _eh.auth_login)
+    events.listen('pytsite.auth.logout', _eh.auth_logout)
+
+    admin_href = router.ep_url('pytsite.odm_ui.ep.browse', {'model': 'auth_log'})
+    admin.sidebar.add_menu('auth', 'auth_log', 'pytsite.auth_log@log', admin_href, 'fa fa-history',
+                           weight=30, permissions='pytsite.odm_ui.browse.auth_log')
+
+__init()