+++ pytsite/content/_widget.py
@@ -3,6 +3,9 @@
 from pytsite import taxonomy as _taxonomy, auth as _auth, widget as _widget, html as _html, lang as _lang, \
     router as _router, tpl as _tpl, odm as _odm
 from . import _model, _functions
+from pytsite import taxonomy as _taxonomy, auth as _auth, widget as _widget, html as _html, lang as _lang, \
+    router as _router, tpl as _tpl, odm as _odm
+from . import _model, _functions
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -13,13 +16,13 @@ class ModelSelect(_widget.select.Select):
     """Content Model Select Widget.
     """
     def __init__(self, **kwargs):
-        self._items = []
+        items = []
         u = _auth.get_current_user()
         for k, v in _functions.get_models().items():
             if u.has_permission('pytsite.odm_ui.browse.' + k) or u.has_permission('pytsite.odm_ui.browse_own.' + k):
-                self._items.append((k, _lang.t(v[1])))
+                items.append((k, _lang.t(v[1])))
 
-        super().__init__(**kwargs)
+        super().__init__(items=sorted(items, key=lambda x: x[1]), **kwargs)
 
 
 class EntitySelect(_widget.select.Select2):